Move as much as simply possible from edge_management into shared code
